Prepare

+++ src/app/Image/imageController.js
@@ -6,6 +6,11 @@ exports.upload = (req, res) => {
 }
 
 exports.delete = (req, res) => {
+    //USE ONLY SERVER
+    if(req.ip != '127.0.0.1' && req.ip != 'localhost') {
+        return res.status(401).json({"message": "Permission denied"})
+    }
+
     fs.unlink(`static/${req.body.filename}`, err => {
         if(err) {
             return res.status(402).json({"message": "Image delete error"});
